+++ packages/icons/CHANGELOG.md
@@ -107,6 +107,12 @@
 - Updated dependencies [de34b15]
   - @marigold/system@18.0.0-beta.0
 
+## 1.3.43
+
+### Patch Changes
+
+- @marigold/system@17.8.0
+
 ## 1.3.42
 
 ### Patch Changes